+++ src/App.js
@@ -2,6 +2,7 @@ import './App.css';
 import MainPage from "./mainpage/mainpage";
 import Signup from './authentication/Signup';
 import Login from './authentication/Login';
+import Profile from './authentication/profile.js';
 import { BrowserRouter as Router} from 'react-router-dom';
 import {Routes, Route} from 'react-router-dom';
 
@@ -14,6 +15,7 @@ function App() {
               <Route path="/" element={<MainPage/>}/>
               <Route path="/signup" element={<Signup/>}/>
               <Route path="/login" element={<Login/>}/>
+              <Route path="/profile" element={<Profile/>}/>
             </Routes>
           </section>
         </div>
